e2e-eval/requirements.txt
coloredlogs
python_on_whales
pyinotify
graphviz
line_profiler
redis
clang==14.0
pyyaml
loguru
psutil
flask
python-dotenv
requests
